reorganize and document
