--- global/global_windows.go
@@ -1,10 +1,0 @@
-package global
-
-import (
-	"github.com/mattn/go-ieproxy"
-	"net/http"
-)
-
-func init() {
-	http.DefaultTransport.(*http.Transport).Proxy = ieproxy.ProxyFromIE
-}
