--- ieproxy_darwin.go
@@ -1,10 +1,0 @@
-// +build !windows
-
-package ieproxy
-
-func GetConf() WindowsProxyConf {
-	return nil
-}
-
-func OverrideEnvWithStaticProxy() {
-}
